Generate Core collections with Length of Encoded Attribute Values

Based on Length of Encoded Attribute Values (LEAV) (Wallace and Boulton 1968; Balakrishnan and Suresh 2001; Balakrishnan and Suresh 2001; Balakrishnan and Nair 2003) estimated from qualitative and/or quantitative trait data, core collections can be generated by the three following methods.

Method I

Classification based on pre-determined diversity represented by LEAV estimates implemented in LEAVcore1.

Method II

Purposive selection of accessions with highest rank of LEAV estimates implemented in LEAVcore2.

Method III

Stratified sampling of accessions from diversity groups/strata computed from LEAV estimates partially implemented in LEAVcore3.

.

Usage

LEAVcore1(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore2(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore3(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)

Arguments

data

The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names

Name of column with the individual names as a character string.

quantitative

Name of columns with the quantitative traits as a character vector.

qualitative

Name of columns with the qualitative traits as a character vector.

size

The desired core set size proportion.

prop.adj

The method for relative frequency transformation for qualitative traits. Either "none" for no transformation or "log" for log-frequency transformation or "sqrt" for square root-proportion transformation (see prop.adj).

e

A named numeric vector of least count of measurement for each quantitative trait specified in quantitative. The list names should be same as quantitative.

Details

Balakrishnan and Suresh (2001); Balakrishnan and Suresh (2001) describe three different methods of constructing core collections from estimates of Length of Encoded Attribute Values.

Method I: Classification based on pre-determined diversity represented by LEAV estimates

This is an objective classification scheme that assigns accessions to either a "core" or "non-core" group based on which group model they best fit.

The target frequency patterns for qualitative traits and distribution parameters for quantitative traits are determined first for the two groups: the Core and the Non-Core.

Target proportions for the core group are estimated from the base proportions of the qualitative trait levels. These may be subjected to transformations if required according to "prop.adj" argument to increase rare trait representation. Target counts are set by scaling these to the total count and capping them at the actual frequency available in the collection. Similarly for the non-core group, the target proportions are determined by subtracting the core model's frequencies from the total counts of each trait level in the entire collection.

The target distribution for the core group is modeled by applying a Gaussian kernel density function to the quantitative trait data, scaled to the core size. The non-core parameters are set to the actual mean and standard deviation of the entire collection.

Based on these target values, the message length (\(F\)) is estimated for each accession against both the models using LEAV. An accession is assigned to the core if \(F_{core} \leq F_{non-core}\). If more accessions are selected than the target core size, the core is refined by ranking individuals by \(F_{core}\) values in ascending order and retaining only the top matches.

Method II: Purposive selection of accessions with highest rank of LEAV estimates

This is a directed selection method that captures the most unique and dispersed accessions to maximize diversity and reduce redundancy.

Here the LEAV index for every accession relative to the entire base collection is first estimated. Then the accessions are ranked in descending order of their LEAV estimates.Finally the core collection is constituted by selecting a pre-determined number of top-ranked accessions according to "size" argument.

Method III: Stratified sampling of accessions from diversity groups/strata computed from LEAV estimates

This is a two-step approach that first organizes the collection into optimized diversity groups based on LEAV estimates followed by a group-wise representative sampling.

Here also the LEAV index for every accession relative to the entire base collection is first estimated. These estimates are then divided into \(L\) strata using the Dalenius formula to minimize pooled variance (Dalenius and Hodges 1959) .

The number of entries to be sampled from each stratum is then determined followed by stratified selection from each group to reach the final core size.

In LEAVcore3, only the stratification based on LEAV estimates is implemented. The downstream steps for allocation (allocate.basic, allocate.diversity, allocate.distance) and stratified selection (select.random, select.diversity, select.distance) are available from the sister package SampleCore.